+++ lib/game.rb
@@ -23,6 +23,7 @@ class Game
   end
 
   def redirect(request)
+    binding.pry
     guess = (request[0].split("=")[1].split(" ")[0]).to_i
     if guess > @value
       #replay game with current @value passed inou
